WS-2876: Refactor [copilot]
