--- src/app/components-webcore/SportDataHeader/head-to-head-v2/helpers/event-summary.ts
@@ -1,225 +1,0 @@
-import { Translations } from '#app/models/types/translations';
-import { EventStatus, HeadToHeadV2Data } from '../types';
-
-const ftFallback = { value: 'FT', accessible: 'Full time' };
-const aetFallback = { value: 'AET', accessible: 'After extra time' };
-
-type PeriodLabel = { value: string; accessible: string; translation?: string };
-
-type RunningScores = {
-  halftime?: string;
-  fulltime?: string;
-  extratime?: string;
-  penaltyShootout?: string;
-  aggregate?: string;
-};
-
-type TeamSummary = {
-  fullName: string;
-  score?: string;
-  runningScores?: RunningScores;
-};
-
-type EventSummaryParams = {
-  time: HeadToHeadV2Data['time'];
-  status: string;
-  periodLabel?: PeriodLabel & { labelType?: string };
-  winner?: string;
-  seriesWinner?: string;
-  home: TeamSummary;
-  away: TeamSummary;
-  period?: string;
-};
-
-type FallbackPeriodLabelParams = {
-  labels: PeriodLabel;
-  status: EventStatus | string;
-  homeRunningScores?: RunningScores;
-  awayRunningScores?: RunningScores;
-  homeName?: string;
-  awayName?: string;
-  translations?: Translations['sport'];
-};
-
-export const getFallbackFootballPeriodLabel = ({
-  labels,
-  status,
-  homeRunningScores,
-  awayRunningScores,
-  homeName,
-  awayName,
-  translations,
-}: FallbackPeriodLabelParams): PeriodLabel => {
-  const isPens = labels?.value?.toLowerCase() === 'pens';
-
-  const {
-    penalties: penaltiesTranslation = 'Penalties',
-    afterExtraTime: aetTranslation,
-    ft: ftTranslation,
-  } = translations || {};
-
-  if (status?.toLowerCase() === 'midevent' && isPens) {
-    const accessibleText = [
-      homeName,
-      ...(homeRunningScores ? [homeRunningScores.penaltyShootout] : []),
-      ',',
-      awayName,
-      ...(awayRunningScores ? [awayRunningScores.penaltyShootout] : []),
-    ].join(' ');
-
-    const hasScore =
-      homeRunningScores?.penaltyShootout && awayRunningScores?.penaltyShootout;
-    const scoreText = hasScore
-      ? ` ${homeRunningScores?.penaltyShootout}-${awayRunningScores?.penaltyShootout}`
-      : '';
-
-    return {
-      value: `${penaltiesTranslation}${scoreText}`,
-      accessible: `${penaltiesTranslation} ${accessibleText}`,
-    };
-  }
-
-  if (isPens) {
-    const aetTranslatedFallback = {
-      ...aetFallback,
-      ...(aetTranslation && {
-        translation: aetTranslation,
-      }),
-    };
-    const ftTranslatedFallback = {
-      ...ftFallback,
-      ...(ftTranslation && {
-        translation: ftTranslation,
-      }),
-    };
-
-    return homeRunningScores?.extratime && awayRunningScores?.extratime
-      ? aetTranslatedFallback
-      : ftTranslatedFallback;
-  }
-
-  return labels;
-};
-
-export const getConciseFootballEventSummary = ({
-  time,
-  status,
-  periodLabel,
-  winner: winnerAlignment,
-  seriesWinner,
-  home,
-  away,
-  period,
-}: EventSummaryParams): string => {
-  const summary: string[] = [];
-  const kickOffTime =
-    time.displayTimeUK !== 'TBC'
-      ? `kick off ${time.accessibleTime}`
-      : time.accessibleTime;
-
-  const homeName =
-    home.fullName === 'TBC' ? 'Team to be confirmed' : home.fullName;
-  const awayName =
-    away.fullName === 'TBC' ? 'Team to be confirmed' : away.fullName;
-
-  if (status === 'PreEvent') {
-    summary.push(homeName, 'versus', awayName, kickOffTime);
-  }
-
-  if (status === 'MidEvent') {
-    summary.push(
-      home.fullName,
-      home.score ?? '',
-      ',',
-      away.fullName,
-      away.score ?? '',
-    );
-
-    if (period === 'et-firsthalf' || period === 'et-secondhalf') {
-      summary.push('Extra time in progress');
-    }
-
-    if (period === 'pens') {
-      const hasRunningScores = home.runningScores && away.runningScores;
-
-      if (hasRunningScores) {
-        if (home.runningScores?.extratime && away.runningScores?.extratime) {
-          summary.push('after extra time');
-        } else {
-          summary.push('after full time');
-        }
-      }
-
-      if (
-        hasRunningScores &&
-        home.runningScores?.penaltyShootout &&
-        away.runningScores?.penaltyShootout
-      ) {
-        summary.push('penalties in progress');
-      } else {
-        summary.push('going to penalties');
-      }
-    }
-  }
-
-  if (status === 'PostEvent') {
-    summary.push(
-      home.fullName,
-      home.score ?? '',
-      ',',
-      away.fullName,
-      away.score ?? '',
-    );
-
-    if (periodLabel?.labelType === 'date') {
-      summary.push('on the');
-    } else if (
-      !home.runningScores?.extratime &&
-      !away.runningScores?.extratime
-    ) {
-      summary.push('at');
-    }
-
-    const fallbackPeriod = getFallbackFootballPeriodLabel({
-      labels: periodLabel ?? ftFallback,
-      status,
-      homeRunningScores: home.runningScores,
-      awayRunningScores: away.runningScores,
-    });
-
-    summary.push(fallbackPeriod.accessible);
-
-    if (
-      home.runningScores?.penaltyShootout &&
-      away.runningScores?.penaltyShootout
-    ) {
-      if (home.runningScores.aggregate && away.runningScores.aggregate) {
-        summary.push(
-          `, ${home.fullName} ${home.runningScores.aggregate} , ${away.fullName} ${away.runningScores.aggregate} on aggregate`,
-        );
-      }
-
-      const { winner, loser } =
-        winnerAlignment?.toLowerCase() === 'home'
-          ? { winner: home, loser: away }
-          : { winner: away, loser: home };
-
-      summary.push(
-        `, ${winner.fullName} win ${winner.runningScores?.penaltyShootout} - ${loser.runningScores?.penaltyShootout} on penalties`,
-      );
-    }
-
-    if (seriesWinner) {
-      const { winner, loser } =
-        seriesWinner.toLowerCase() === 'home'
-          ? { winner: home, loser: away }
-          : { winner: away, loser: home };
-
-      summary.push(
-        `, ${winner.fullName} win ${winner.runningScores?.aggregate} - ${loser.runningScores?.aggregate} on aggregate`,
-      );
-    }
-  }
-
-  return summary.join(' ');
-};
